Move functions from FrontPage.js to util.js
