+++ src/components/CourseList.js
@@ -2,9 +2,12 @@ import React, {PropTypes} from 'react';
 import CourseItem from './CourseItem';
 
 const CourseList = React.createClass({
-
+  sortCourses(a, b) {
+    return b.lessons.length - a.lessons.length;
+  },
   render() {
-    const courses = this.props.courses.map((course, idx) => {
+    const sortedCourses = this.props.courses.sort(this.sortCourses);
+    const courses = sortedCourses.map((course, idx) => {
       if (course.lessons.length == 0) return null;
       return (
         <CourseItem key={idx} course={course}/>
